Safe to run twice, null _parent, 1.0.2
There was a problem that if you run this over the same data set over and
over, additional `propName?` and `propName??` and `propName???`
properties would be set.  Fixed.

Also, file objects should have had a `null` `_parent` property, but they
would not do that if they linked to each other in some sort of tree.
Fixed.

Releasing 1.0.2.